mvc stracture

+++ src/views/invoice/ilist.php
@@ -66,7 +66,7 @@
                                 <a href="edit.php?inv_number=<?php echo $row['inv_number'] ?>" class="link-dark">
                                     <i class="fa-solid fa-pen-to-square fs-5 me-3"></i>
                                 </a>
-                                <a href="../../classes/invoice/deleteInvoice.php?inv_number=<?php echo $row['inv_number'] ?>" class="link-dark">
+                                <a href="../../classes/invoice/deleteInvoice.php?inv_number=<?php echo $row['inv_number'] ?>" class="link-dark" title="Delete" onclick="return confirm('Are you sure you want to delete this employee?');">
                                     <i class="fa-solid fa-trash fs-5"></i>
                                 </a>
                             </td>
